AMD FSR 2.2 implementation
This is still missing the correct delta time passed into the AMD FSR 2.2 dispatch, as well as the sharpen factor.
The options and debug menu have been adjusted, now you can select a "Upscaling and Anti-aliasing" method, which are TAA, FSR 2.2, and Off.
Release and ship builds to be tested soon.